rename test file and use the openskillrating object in pufferlib rather than a copy/paste
